requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
dj-database-url==2.1.0
Django==5.0.6
django-ckeditor==6.7.1
django-countries==7.6.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
